// src/views/Vendors/types.ts
/**
 * Vendor type definitions
 * @module sbom-harbor-ui/views/Vendors/types
 */

export interface Vendor {
  id: string
  name: string
  description: string
  contact: string
  email: string
  phone: string
  address: string
  city: string
  state: string
  zip: string
  country: string
  website: string
  notes: string
}